[#54] refactor: SoftDeleteEntity 도입 및 엔티티 구조 개선
