--- src/main/java/com/creatorhub/entity/BaseEntity.java
@@ -1,22 +1,0 @@
-package com.creatorhub.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@EntityListeners(value = {AuditingEntityListener.class})
-@MappedSuperclass
-@Getter
-public abstract class BaseEntity extends BaseTimeEntity {
-    @CreatedBy
-    @Column(updatable = false, length = 100)
-    private String createdBy;
-
-    @LastModifiedBy
-    @Column(length = 100)
-    private String updatedBy;
-}
