--- src/main/java/com/creatorhub/entity/BaseTimeEntity.java
@@ -1,36 +1,0 @@
-package com.creatorhub.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@EntityListeners(value = {AuditingEntityListener.class})
-@MappedSuperclass
-@Getter
-public abstract class BaseTimeEntity {
-    @CreatedDate
-    @Column(updatable = false, nullable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(nullable = false)
-    private LocalDateTime updatedAt;
-
-    private LocalDateTime deletedAt;
-
-    // 삭제 데이터 복구
-    public void restore() {
-        this.deletedAt = null;
-    }
-
-    // 삭제 여부 확인
-    public boolean isDeleted() {
-        return deletedAt != null;
-    }
-}
