+++ src/main/java/com/creatorhub/entity/base/BaseSoftDeleteEntity.java
@@ -1,0 +1,15 @@
+package com.creatorhub.entity.base;
+
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@EntityListeners(value = {AuditingEntityListener.class})
+@MappedSuperclass
+@Getter
+public abstract class BaseSoftDeleteEntity extends BaseEntity {
+    private LocalDateTime deletedAt;
+}
